First functional prototype
* Added askForMasterPwd() to ioCLI
* Added readChNoEcho() to ioCLI
* Made ioCLI more compact with readStr(), readLong(), and invalidNatNum()
* Increased the largest displayable line number to 99999 (now uses long type)
* Improved error messages and made eraseParams() precede exit() calls

+++ inc/ioFile.h
@@ -5,6 +5,6 @@
 #define loadAccountLine(a) pwdgenLoadAccountLine(a)
 
 void saveAccount();
-int loadAccountLine(unsigned lineNum);
+int loadAccountLine(long lineNum);
 
 #endif
